+++ index.js
@@ -234,6 +234,8 @@ module.exports = {
     'no-delete-var': 1,
     // disallow labels that share a name with a variable
     'no-label-var': 1,
+    // restrict usage of specified global variables
+    'no-restricted-globals': 0,
     // disallow shadowing of names such as arguments
     'no-shadow-restricted-names': 1,
     // disallow declaration of variables already declared in the outer scope
@@ -336,6 +338,8 @@ module.exports = {
     'new-parens': 1,
     // require or disallow an empty newline after variable declarations
     'newline-after-var': 1,
+    // require newline before return statement
+    'newline-before-return': 0,
     //  enforce newline after each call when chaining the calls
     'newline-per-chained-call': 0,
     // disallow use of the Array constructor
@@ -463,6 +467,8 @@ module.exports = {
     'prefer-template': 1,
     // disallow generator functions that do not have yield
     'require-yield': 1,
+    // enforce spacing around embedded expressions of template strings
+    'template-curly-spacing': 1,
     // enforce spacing around the * in yield* expressions
     'yield-star-spacing': 1,